Add tagListerAtoZ snippet with multi TV support that lists tags separated in alphabetical groups.
Abstracts out the actual retrieving of tags to a class method which could also be used for the main tagLister snippet.
Kinda tight on time to also make the main tagLister snippet use it right now as it would need some refactoring to use the method, but it's on the ever-growing list of things to do.